+++ src/Tools/dotnet-monitor/CollectionRules/Actions/CollectLogsAction.cs
@@ -5,6 +5,7 @@
 using Microsoft.Diagnostics.Monitoring.EventPipe;
 using Microsoft.Diagnostics.Monitoring.Options;
 using Microsoft.Diagnostics.Monitoring.WebApi;
+using Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Exceptions;
 using Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Options.Actions;
 using Microsoft.Extensions.Logging;
 using System;
@@ -83,7 +84,10 @@ namespace Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Actions
                     scope);
 
                 ExecutionResult<EgressResult> result = await egressOperation.ExecuteAsync(_serviceProvider, token);
-
+                if (null != result.Exception)
+                {
+                    throw new CollectionRuleActionException(result.Exception);
+                }
                 string logsFilePath = result.Result.Value;
 
                 return new CollectionRuleActionResult()