+++ documentation/releaseNotes/releaseNotes.md
@@ -1,4 +1,3 @@
 Today we are releasing the next official preview of the `dotnet monitor` tool. This release includes:
 
-- Add experience survey link (#1601)
-- Docker: Add CBL-Mariner 2.0 amd64 image
+- Added Collection Rule Defaults (#1595)